tests: fix handling of tmp dir
If the tests are run in a fresh git checkout where the tests/data/tmp
directory doesn't exist, the test_with_output_dir() case would create it
as a *file* instead of a directory, and then the test_with_output_file()
case would subsequently fail.

Ensure the tmp directory is actually a directory in each case.

Also ensure the tmp directory itself is removed at the end of the tests,
which reveals this issue.